+++ caffe2/operators/sigmoid_op.cc
@@ -1,10 +1,5 @@
 #include "caffe2/operators/sigmoid_op.h"
 
-#include <algorithm>
-#include <functional>
-
-#include "caffe2/utils/math.h"
-
 namespace caffe2 {
 
 template <>
@@ -16,42 +11,12 @@ operator()(const int N, const T* X, T* Y, CPUContext* /* context */) const {
   return true;
 }
 
-#if !CAFFE2_MOBILE
-
-template <>
-template <typename T>
-bool SigmoidGradientFunctor<CPUContext>::Forward(
-    const std::vector<int>& dY_dims,
-    const std::vector<int>& /* Y_dims */,
-    const T* dY,
-    const T* Y,
-    T* dX,
-    CPUContext* /* context */) const {
-  const int size = std::accumulate(
-      dY_dims.cbegin(), dY_dims.cend(), 1, std::multiplies<int>());
-  ConstEigenVectorArrayMap<T> dY_arr(dY, size);
-  ConstEigenVectorArrayMap<T> Y_arr(Y, size);
-  EigenVectorArrayMap<T>(dX, size) = dY_arr * Y_arr * (1. - Y_arr);
-  return true;
-}
-
-#endif // !CAFFE2_MOBILE
-
 REGISTER_CPU_OPERATOR(
     Sigmoid,
     UnaryElementwiseOp<
         TensorTypes<float>,
         CPUContext,
         SigmoidFunctor<CPUContext>>);
-
-#if !CAFFE2_MOBILE
-
-REGISTER_CPU_OPERATOR(
-    SigmoidGradient,
-    BinaryElementwiseOp<
-        TensorTypes<float>,
-        CPUContext,
-        SigmoidGradientFunctor<CPUContext>>);
 
 // Input: X, output: Y
 OPERATOR_SCHEMA(Sigmoid)
@@ -76,24 +41,5 @@ OPERATOR_SCHEMA(SigmoidGradient)
 SigmoidGradient takes both Y and dY and uses this to update dX according to the
 chain rule and derivatives of the sigmoid function.
 )DOC");
-
-namespace {
-
-class GetSigmoidGradient : public GradientMakerBase {
-  using GradientMakerBase::GradientMakerBase;
-  std::vector<OperatorDef> GetGradientDefs() override {
-    return SingleGradientDef(
-        "SigmoidGradient",
-        "",
-        std::vector<std::string>{GO(0), O(0)},
-        std::vector<std::string>{GI(0)});
-  }
-};
-
-} // namespace
-
-REGISTER_GRADIENT(Sigmoid, GetSigmoidGradient);
-
-#endif // !CAFFE2_MOBILE
 
 } // namespace caffe2
